mobile-dev/backend/requirements.txt
fastapi==0.101.1
uvicorn[standard]==0.22.0
requests==2.31.0
python-dotenv==1.0.0
